Creation and setting up the project. Added:  entities;  Role, User Repository;  Role, User Service;  Main and Registration Controller;  html and css files for registration

+++ src/main/java/com/example/demo/entities/Role.java
@@ -17,7 +17,7 @@ public class Role {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Id
     @Column(name = "id", nullable = false)
-    private int id;
+    private Long id;
     @Basic
     @Column(name = "name", nullable = false, length = -1)
     private String name;
@@ -25,7 +25,7 @@ public class Role {
     @OneToMany(mappedBy = "roleId")
     private Collection<User> usersId;
 
-    public Role(int id, String name) {
+    public Role(Long id, String name) {
         this.id = id;
         this.name = name;
     }
